requirements.txt
django~=4.1
djangorestframework~=3.15.2
pillow~=10.4.0
markdown~=3.7
django-filter~=24.3
python-dotenv~=1.0.1
psycopg2-binary~=2.9.9
drf-yasg~=1.21.7
djangorestframework-simplejwt~=5.3.1
django-cors-headers~=4.4.0
django_pghistory~=3.4.0
django_pgtrigger~=4.12